+++ app/Harvest/Mappers/CollectionItemMapper.php
@@ -4,12 +4,11 @@ namespace App\Harvest\Mappers;
 
 use App\Collection;
 
-class CollectionMapper extends AbstractModelMapper
+class CollectionItemMapper extends AbstractModelMapper
 {
     protected $modelClass = Collection::class;
 
     public function mapId(array $row) {
         return $row['id'];
     }
-
 }